# server/swagger.yaml
swagger: '2.0'
info:
  description: >-
    This is an API developed for a fleet monitoring system software. This is
    used to allow communication from microcontrollers and the website to the
    server..
  version: 1.0.0
  title: Fleet Monitoring System
  contact:
    email: sdmay18-18@iastate.edu
host: '35.193.87.37:8080'
...
schemes:
  - http
paths:
  /vehicle:
    post:
      tags:
        - vehicle
      summary: Add a new vehicle
      description: ''
      operationId: addVehicle
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: uid
          description: The uid of the vehicle to create.
          schema:
            type: object
            required:
              - uid
            properties:
              uid:
                type: number
      responses:
        '405':
          description: Invalid input
    get:
      tags:
        - vehicle
      summary: List all vehicles
      description: A listAll for vehicles. This is mainly used for testing.
      operationId: getVehicles
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Vehicle'
        '400':
          description: Invalid status value
  '/vehicle/{uid}':
    get:
      tags:
        - vehicle
      summary: Finds Vehicle by uid
      description: Gets data of a specific vehicle
      operationId: findVehicleByUID
      produces:
        - application/json
      parameters:
        - name: uid
          in: path
          description: uid of the vehicle
          required: true
          type: number
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Vehicle'
        '400':
          description: Invalid uid
    put:
      tags:
        - vehicle
      summary: Updated vehicle data
      description: Updated vehicle data from microcontroller.
      operationId: updateVehicle
      produces:
        - application/json
      parameters:
        - name: uid
          in: path
          description: uid of the vehicle that needs to be updated
          required: true
          type: number
        - in: body
          name: body
          description: Updated vehicle data object
          required: true
          schema:
            $ref: '#/definitions/VehicleData'
      responses:
        '400':
          description: Invalid data supplied
        '404':
          description: Vehicle not found
    delete:
      tags:
        - vehicle
      summary: Delete vehicle
      description: Deletes a vehicle by uid.
      operationId: deleteVehcile
      produces:
        - application/json
      parameters:
        - name: uid
          in: path
          description: The uid of the vehicle that needs to be deleted
          required: true
          type: string
      responses:
        '400':
          description: Invalid uid supplied
        '404':
          description: Vehicle not found
...
definitions:
  Vehicle:
    type: object
    properties:
      uid:
        type: number
      mrGas:
        type: number
      mrSpeed:
        type: number
      mrLong:
        type: number
      mrLat:
        type: number
      data:
        type: array
        items:
          $ref: '#/definitions/VehicleData'
...
  VehicleData:
    type: object
    properties:
      speed:
        type: number
      gas:
        type: number
      longitude:
        type: number
      latitude:
        type: number
      date:
        type: string
...